fix: align storage metadata GPRC bind port with other variable names
Currently STORAGE_METADATA_GRPC_PROVIDER_ADDR environment variable used as the way to set the GRPC bind port for the service.  All other services use the format <SERVICE_NAME>_GRPC_ADDR for GRPC, and <SERVICE_NAME>_HTTP_ADDR for HTTP.  This change brings it in line with the others, as it's confusing why this one is different.